requirements-base.txt
coreapi==2.3.3
Django==3.1.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
dj-rest-auth==2.1.4
django-allauth==0.44.0
django_compressor==2.4
django-cors-headers==3.7.0
django-countries==7.1
django-crispy-forms==1.11.2
django-environ==0.4.5
django-libsass==0.8
django-pwa==1.0.10
django-rest-swagger==2.2.0
django-sendgrid-v5==0.9.1
django_storages[azure]==1.11.1
pyTelegramBotAPI==3.7.7
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
requests==2.25.1
